requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Pillow==9.5.0
qrcode==7.3
reportlab==4.0.10
gunicorn==20.1.0
